# docs/api-docs/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "中间层模块",
        "description": "虚拟宠物医院API文档，前端用。字段默认值：int为-1，String为'This field is currently Empty.'，Date为null，应该会表现为空字符串。",
        "version": "1.0.1-SNAPSHOT"
    },
    "servers": [
        {
            "url": "http://pikapikpinoimac:8090",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/system/examine": {
            "get": {
                "tags": [
                    "系统管理模块"
                ],
                "summary": "获取化验信息列表接口",
                "operationId": "getExamines",
                "parameters": [
                    {
                        "name": "examinePrice",
                        "in": "query",
                        "description": "化验价格",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    },
                    {
                        "name": "roomName",
                        "in": "query",
                        "description": "科室名称",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "515": {
                        "description": "Request failed",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 515,
                                    "message": "Request failed"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 200,
                                    "data": {
                                        "examineList": [
                                            {
                                                "examineName": "testroom",
                                                "examineRoom": {
                                                    "roomRole": "this is a test",
                                                    "roomName": "testroom"
                                                },
                                                "examineId": 2,
                                                "examinePrice": 1.2
                                            },
                                            {
                                                "examineName": "Examine Item 4",
                                                "examineRoom": {
                                                    "roomRole": "RAD",
                                                    "roomName": "examine room"
                                                },
                                                "examineId": 3,
                                                "examinePrice": 13.12
                                            }
                                        ]
                                    },
                                    "message": "ok"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "系统管理模块"
                ],
                "summary": "更改化验信息接口",
                "operationId": "updateExamine",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "json",
                                "description": "更改后的化验信息",
                                "example": {
                                    "examineId": 2,
                                    "examineName": "Examine Item 1",
                                    "examinePrice": 123.12,
                                    "roomName": "examine room"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 200,
                                    "data": {
                                        "examineName": "Examine Item 4",
                                        "examineRoom": {
                                            "roomRole": "RAD",
                                            "roomName": "examine room"
                                        },
                                        "examineId": 3,
                                        "examinePrice": 123.12
                                    },
                                    "message": "ok"
                                }
                            }
                        }
                    },
                    "515": {
                        "description": "Request failed",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 515,
                                    "message": "Request failed"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "系统管理模块"
                ],
                "summary": "新增化验信息接口",
                "operationId": "addExamine",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "json",
                                "description": "新增的化验信息",
                                "example": {
                                    "examineName": "Examine Item 1",
                                    "examinePrice": 123.12,
                                    "roomName": "examine room"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "515": {
                        "description": "Request failed",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 515,
                                    "message": "Request failed"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 200,
                                    "data": {
                                        "examineName": "Examine Item 4",
                                        "examineRoom": {
                                            "roomRole": "RAD",
                                            "roomName": "examine room"
                                        },
                                        "examineId": 3,
                                        "examinePrice": 13.12
                                    },
                                    "message": "ok"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/system/examine/{examineId}": {
            "delete": {
                "tags": [
                    "系统管理模块"
                ],
                "summary": "删除化验信息接口",
                "operationId": "deleteExamine",
                "parameters": [
                    {
                        "name": "examineId",
                        "in": "path",
                        "description": "化验信息Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 200,
                                    "data": {
                                        "examineName": "testroom",
                                        "examineId": 2,
                                        "examinePrice": 1.2
                                    },
                                    "message": "ok"
                                }
                            }
                        }
                    },
                    "515": {
                        "description": "Request failed",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 515,
                                    "message": "Request failed"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}